# scdms-backend/src/assets/OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCDMS - Simplified Coding Data Managment Service",
    "description": "This is the Simplified Coding Data Managment Service API.",
    "contact": {
      "email": "main@simplifiecoding.org"
    },
    "license": {
      "name": "GPL v3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "https://scdms-server.simplifiedcoding.org"
    }
  ],
...
  "paths": {
...
    "/oauth/discord/state": {
      "post": {
        "operationId": "post_oauth_discord_state",
        "summary": "POSTs to the server a state",
        "description": "Appends to the states object a new state ID to allow external authentication",
        "tags": [
          "oauth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "state"
                ],
                "properties": {
                  "state": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "default": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/oauth/discord/finalize": {
      "get": {
        "operationId": "get_oauth_discord_finalize",
        "summary": "Finalize a discord OAuth request & get access token",
        "description": "Redirects to discord OAuth authorize URL.",
        "tags": [
          "oauth"
        ],
        "responses": {
          "200": {
            "description": "Account authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "default": true
                    },
                    "state": {
                      "type": "string"
                    },
                    "jwt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found / Incorrent Params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "default": true
                    },
                    "msg": {
                      "type": "string",
                      "example": "The state was invalid or the IP address doesn't match with the original one"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User JWT"
      }
    }
  }
}